refactor(access): move access packge into remote directory
this clarifies the pacakge is only about access in relation to remotes
